direct login resolve redudant
