+++ web/app/signin/page.tsx
@@ -1,32 +1,22 @@
+import type { SearchParams } from '@/utils/search-params'
 import { cookies } from 'next/headers'
 import { redirect } from 'next/navigation'
+import { firstSearchParam, searchParamsToString } from '@/utils/search-params'
 import SignInClient from './signin-client'
 
-type SearchParams = Record<string, string | string[] | undefined>
 type CookieStore = Awaited<ReturnType<typeof cookies>>
 
-const first = (value: string | string[] | undefined) => Array.isArray(value) ? value[0] : value
-
-const toSearchString = (searchParams: SearchParams) => {
-  const params = new URLSearchParams()
-  for (const [key, value] of Object.entries(searchParams)) {
-    if (typeof value === 'string')
-      params.set(key, value)
-    else if (Array.isArray(value))
-      value.forEach(v => params.append(key, v))
-  }
-  const text = params.toString()
-  return text ? `?${text}` : ''
-}
+const hasConsoleAccessTokenCookie = (cookieStore: CookieStore) =>
+  Boolean(cookieStore.get('access_token')?.value || cookieStore.get('__Host-access_token')?.value)
 
 const shouldAutoRedirectToAceDataCloudOAuth = (cookieStore: CookieStore, searchParams: SearchParams) => {
-  if (cookieStore.get('access_token')?.value)
+  if (hasConsoleAccessTokenCookie(cookieStore))
     return false
-  if (first(searchParams.message))
+  if (firstSearchParam(searchParams.message))
     return false
-  if (first(searchParams.step) === 'next')
+  if (firstSearchParam(searchParams.step) === 'next')
     return false
-  if (first(searchParams.no_acedatacloud_oauth) === '1')
+  if (firstSearchParam(searchParams.no_acedatacloud_oauth) === '1')
     return false
   return true
 }
@@ -35,7 +25,7 @@ const SignIn = async ({ searchParams }: { searchParams?: SearchParams }) => {
   const safeSearchParams = searchParams || {}
   const cookieStore = await cookies()
   if (shouldAutoRedirectToAceDataCloudOAuth(cookieStore, safeSearchParams)) {
-    const query = toSearchString(safeSearchParams)
+    const query = searchParamsToString(safeSearchParams)
     redirect(`/console/api/oauth/login/acedatacloud${query}`)
   }
   return <SignInClient />
